+++ 3-rest-api/controllers/getAll.js
@@ -1,4 +1,4 @@
-const Contact = require("../models/contacts");
+const { Contact } = require("../models/contacts");
 
 const getAll = async (req, res, next) => {
   try {
