+++ 3-rest-api/controllers/getById.js
@@ -1,4 +1,4 @@
-const Contact = require("../models/contacts");
+const { Contact } = require("../models/contacts");
 const { createError } = require("../helpers");
 
 const getById = async (req, res, next) => {
