+++ 3-rest-api/routes/api/contacts.js
@@ -1,6 +1,7 @@
 const express = require("express");
 
 const ctrl = require("../../controllers");
+
 const router = express.Router();
 
 router.get("/", ctrl.getAll);
